重新拆分lib.element
