+++ noname/util/index.js
@@ -14,8 +14,8 @@ export class Uninstantable {
 
 /**
  * 
- * @param {number} ms 
- * @returns {Promise<void>}
+ * @param { number } ms 
+ * @returns { Promise<void> }
  */
 export function delay(ms) {
 	return new Promise((resolve) => setTimeout(resolve, ms));
